Reduce realtime logging noise
