+++ packages/sdk/src/utils/media.ts
@@ -1,0 +1,49 @@
+export async function blobToBase64(blob: Blob): Promise<string> {
+  return new Promise((resolve, reject) => {
+    const reader = new FileReader();
+    reader.onloadend = () => {
+      const result = reader.result;
+      if (typeof result !== "string") {
+        reject(new Error("FileReader did not return a string"));
+        return;
+      }
+      const base64 = result.split(",")[1];
+      if (!base64) {
+        reject(new Error("Invalid data URL format"));
+        return;
+      }
+      resolve(base64);
+    };
+    reader.onerror = reject;
+    reader.readAsDataURL(blob);
+  });
+}
+
+export async function imageToBase64(image: Blob | File | string): Promise<string> {
+  if (typeof image === "string") {
+    let url: URL | null = null;
+    try {
+      url = new URL(image);
+    } catch {
+      // Not a valid URL, treat as raw base64
+    }
+
+    if (url?.protocol === "data:") {
+      const [, base64] = image.split(",", 2);
+      if (!base64) {
+        throw new Error("Invalid data URL image");
+      }
+      return base64;
+    }
+    if (url?.protocol === "http:" || url?.protocol === "https:") {
+      const response = await fetch(image);
+      if (!response.ok) {
+        throw new Error(`Failed to fetch image: ${response.status} ${response.statusText}`);
+      }
+      const imageBlob = await response.blob();
+      return blobToBase64(imageBlob);
+    }
+    return image;
+  }
+  return blobToBase64(image);
+}
